Work in progress - C# API.
git-svn-id: https://vdsf.svn.sourceforge.net/svnroot/vdsf/trunk@900 53c26a9f-4021-0410-9f73-c4284c400f72

+++ src/CS/PhotonLib/PrivateFolder.cs
@@ -1,3 +1,21 @@
+/*
+ * Copyright (C) 2006-2008 Daniel Prevost <dprevost@photonsoftware.org>
+ *
+ * This file is part of Photon (photonsoftware.org).
+ *
+ * This file may be distributed and/or modified under the terms of the
+ * GNU General Public License version 2 or version 3 as published by the 
+ * Free Software Foundation and appearing in the file COPYING.GPL2 and 
+ * COPYING.GPL3 included in the packaging of this software.
+ *
+ * Licensees holding a valid Photon Commercial license can use this file 
+ * in accordance with the terms of their license.
+ *
+ * This software is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. 
+ */
+
 using System;
 using System.Collections.Generic;
 using System.Text;
@@ -22,20 +40,20 @@ namespace Photon
         private static extern int psoFolderCreateObject( 
             IntPtr           folderHandle,
             string           objectName,
-            IntPtr           nameLengthInBytes,
+            UInt32           nameLengthInBytes,
             ObjectDefinition pDefinition);
 
         [DllImport("photon.dll", EntryPoint = "psoFolderCreateObjectXML", CallingConvention = CallingConvention.Cdecl)]
         private static extern int psoFolderCreateObjectXML( 
             IntPtr folderHandle,
             string xmlBuffer,
-            IntPtr lengthInBytes );
+            UInt32 lengthInBytes );
 
         [DllImport("photon.dll", EntryPoint = "psoFolderDestroyObject", CallingConvention = CallingConvention.Cdecl)]
         private static extern int psoFolderDestroyObject( 
             IntPtr folderHandle,
             string objectName,
-            IntPtr nameLengthInBytes );
+            UInt32 nameLengthInBytes );
 
         [DllImport("photon.dll", EntryPoint = "psoFolderGetFirst", CallingConvention = CallingConvention.Cdecl)]
         private static extern int psoFolderGetFirst( 
@@ -51,7 +69,7 @@ namespace Photon
         private static extern int psoFolderOpen( 
             IntPtr     sessionHandle,
             string     folderName,
-            IntPtr     nameLengthInBytes,
+            UInt32     nameLengthInBytes,
             ref IntPtr objectHandle );
 
         [DllImport("photon.dll", EntryPoint = "psoFolderStatus", CallingConvention = CallingConvention.Cdecl)]
